+++ src/ts/store.ts
@@ -48,7 +48,7 @@ namespace CdvPurchase {
     /**
      * Current release number of the plugin.
      */
-    export const PLUGIN_VERSION = '13.18.0';
+    export const PLUGIN_VERSION = '13.18.1';
 
     /**
      * Entry class of the plugin.